Eliminate warnings related to conversion from size_t to unsigned int in cuda::reduce_n Eliminate warnings related to empty switch statement on Windows in error_category

+++ thrust/detail/device/cuda/reduce_n.inl
@@ -148,11 +148,11 @@ template<typename InputIterator,
                       thrust::detail::true_type)    // reduce in shared memory
 {
     // determine launch parameters
-    const size_t smem_per_thread = sizeof(OutputType);
-    const size_t block_size = thrust::experimental::arch::max_blocksize_with_highest_occupancy(reduce_n_smem<InputIterator, OutputType, BinaryFunction>, smem_per_thread);
-    const size_t smem_size  = block_size * smem_per_thread;
-    const size_t max_blocks = thrust::experimental::arch::max_active_blocks(reduce_n_smem<InputIterator, OutputType, BinaryFunction>, block_size, smem_size);
-    const size_t num_blocks = std::min(max_blocks, (n + (block_size - 1)) / block_size);
+    const unsigned int smem_per_thread = sizeof(OutputType);
+    const unsigned int block_size = thrust::experimental::arch::max_blocksize_with_highest_occupancy(reduce_n_smem<InputIterator, OutputType, BinaryFunction>, smem_per_thread);
+    const unsigned int smem_size  = block_size * smem_per_thread;
+    const unsigned int max_blocks = thrust::experimental::arch::max_active_blocks(reduce_n_smem<InputIterator, OutputType, BinaryFunction>, block_size, smem_size);
+    const unsigned int num_blocks = std::min(max_blocks, (n + (block_size - 1)) / block_size);
 
     // allocate storage for per-block results
     thrust::detail::raw_cuda_device_buffer<OutputType> temp(num_blocks + 1);
@@ -189,11 +189,11 @@ template<typename InputIterator,
                       thrust::detail::false_type)    // reduce in global memory
 {
     // determine launch parameters
-    const size_t smem_per_thread = 0;
-    const size_t block_size = thrust::experimental::arch::max_blocksize_with_highest_occupancy(reduce_n_gmem<InputIterator, OutputType, BinaryFunction>, smem_per_thread);
-    const size_t smem_size  = block_size * smem_per_thread;
-    const size_t max_blocks = thrust::experimental::arch::max_active_blocks(reduce_n_gmem<InputIterator, OutputType, BinaryFunction>, block_size, smem_size);
-    const size_t num_blocks = std::min(max_blocks, (n + (block_size - 1)) / block_size);
+    const unsigned int smem_per_thread = 0;
+    const unsigned int block_size = thrust::experimental::arch::max_blocksize_with_highest_occupancy(reduce_n_gmem<InputIterator, OutputType, BinaryFunction>, smem_per_thread);
+    const unsigned int smem_size  = block_size * smem_per_thread;
+    const unsigned int max_blocks = thrust::experimental::arch::max_active_blocks(reduce_n_gmem<InputIterator, OutputType, BinaryFunction>, block_size, smem_size);
+    const unsigned int num_blocks = std::min(max_blocks, (n + (block_size - 1)) / block_size);
 
     // allocate storage for per-block results
     thrust::detail::raw_cuda_device_buffer<OutputType> temp(num_blocks + 1);
